Changed the level again and added rock 'n' roll:
- the wall jump "zone" consists now of several cubes forming two bridge like structures (the new material on the cubes is not final)
- added a mountain with a flag on top and 30 rocks that spawn when the player gets near
- added a position text to the text output at the top which makes it much easier to place flags. Improved flag positions.

The current rock model is completely made in blender (with backed materials). The material looks quite interesting, but does totally not fit to the mountain. The terrain has no normal map, that would improve a lot.
The mountain with the rolling rocks is not yet as I imagined it. I'll change it again to a bigger mountain with a less steep slope.
There are currently several issues: Like the player position not resetting when restarting the level and the rocks only spawning after first entering the level. Also bad code structure.
The physics have a weird frame rate issue: With 30 or 60 rocks I have constantly over 60 FPS. But with 100 rocks I have less than 10 FPS when the rocks hit the ground and that lasts for several seconds.